Move CLI service registry from plugin to CLI
Relocate the CLI service registry from src/plugin/src/cli-services/ into src/cli/lib/plugin-service-providers so the CLI package owns its wiring instead of the formatter package exposing CLI-only APIs.
