--- src/plugin/src/cli-services.js
@@ -1,15 +1,0 @@
-export {
-    createDefaultCliPluginServices,
-    createDefaultCliProjectIndexService,
-    createDefaultCliIdentifierCasePlanService,
-    createDefaultCliIdentifierCasePlanPreparationService,
-    createDefaultCliIdentifierCaseCacheService,
-    resolveCliPluginServices,
-    resolveCliProjectIndexService,
-    resolveCliIdentifierCasePlanService,
-    resolveCliIdentifierCasePlanPreparationService,
-    resolveCliIdentifierCaseCacheService,
-    defaultIdentifierCasePlanPreparer,
-    defaultProjectIndexBuilder,
-    defaultIdentifierCaseCacheClearer
-} from "./cli-services/cli-plugin-service-registry.js";
