+++ src/plugin/src/identifier-case.js
@@ -1,0 +1,3 @@
+export { prepareIdentifierCasePlan } from "./identifier-case/plan-service.js";
+export { clearIdentifierCaseOptionStore } from "./identifier-case/option-store.js";
+export { clearIdentifierCaseDryRunContexts } from "./identifier-case/identifier-case-context.js";
