+++ src/plugin/src/project-index.js
@@ -1,0 +1,3 @@
+export { buildProjectIndex } from "./project-index/index.js";
+export { getProjectIndexParserOverride } from "./project-index/index.js";
+export { getDefaultFsFacade } from "./project-index/index.js";
